+++ subs/hweight.f
@@ -23,10 +23,10 @@ C                     assume X(N+1)=0 with W(N+1)=0.0 with out loss of
 C                     accuracy.
 C Created 25-Nov-1986 (Based on KWEIGHT)
 C
-	INTEGER*4 N,I
+	INTEGER N,I
 	REAL*8 X(N),W(N),H,HN,RF,RE,SUM
 C
-	INTEGER*4 LUER,ERROR_LU
+	INTEGER LUER,ERROR_LU
 	EXTERNAL ERROR_LU
 C
 	LUER=ERROR_LU()
