+++ subs/locsolut.f
@@ -4,10 +4,10 @@ C
 C
 C Altered 24-May-1996 : DOUBLE PRECISION declaration removed.
 C
-	INTEGER*4 ND,NT,LU
+	INTEGER ND,NT,LU
 	REAL*8 STEQ(NT,ND),POPS(NT,ND),TA(ND)
 C
-	INTEGER*4 I,J,K,M,MIN,MAX
+	INTEGER I,J,K,M,MIN,MAX
 	REAL*8 T1,T2,T3,SCALE
 C
 C Read in solution matrix (STEQ) , and scaling vector (TA).
